[FEAT] : 3015 오아시스 재결합
N이 최대 50만개라 pair는 long으로...

+++ 05_Stack/problem 3015/src/Main.java
@@ -13,7 +13,7 @@ public class Main {
         int N = Integer.parseInt(br.readLine());
 
         // 쌍의 수
-        int pair = 0;
+        long pair = 0;
 
         for (int i = 0; i < N; i++) {
             int num = Integer.parseInt(br.readLine());
